+++ metadata.json
@@ -1,1 +1,1 @@
-{ "DEFAULT_FEE_RECIPIENT": "" }
+{"DEFAULT_FEE_RECIPIENT":""}
