+++ ts/src/middleware/error_handling.ts
@@ -3,11 +3,57 @@ import * as HttpStatus from 'http-status-codes';
 
 import {
     BadRequestError,
+    ErrorBodyWithHTTPStatusCode,
     GeneralErrorCodes,
     generalErrorCodeToReason,
     RelayerBaseError,
     ValidationError,
 } from '../errors';
+
+/**
+ * Wraps an Error with a JSON human readable reason and status code.
+ */
+export function generateError(err: Error): ErrorBodyWithHTTPStatusCode {
+    if ((err as any).isRelayerError) {
+        const relayerError = err as RelayerBaseError;
+        const statusCode = relayerError.statusCode;
+        if (relayerError.statusCode === HttpStatus.BAD_REQUEST) {
+            const badRequestError = relayerError as BadRequestError;
+            if (badRequestError.generalErrorCode === GeneralErrorCodes.ValidationError) {
+                const validationError = badRequestError as ValidationError;
+                return {
+                    statusCode,
+                    errorBody: {
+                        code: badRequestError.generalErrorCode,
+                        reason: generalErrorCodeToReason[badRequestError.generalErrorCode],
+                        validationErrors: validationError.validationErrors,
+                    },
+                };
+            } else if (badRequestError.generalErrorCode === GeneralErrorCodes.MalformedJson) {
+                return {
+                    statusCode,
+                    errorBody: {
+                        code: badRequestError.generalErrorCode,
+                        reason: generalErrorCodeToReason[badRequestError.generalErrorCode],
+                    },
+                };
+            }
+        } else {
+            return {
+                statusCode,
+                errorBody: {
+                    reason: HttpStatus.getStatusText(relayerError.statusCode),
+                },
+            };
+        }
+    }
+    return {
+        statusCode: HttpStatus.BAD_REQUEST,
+        errorBody: {
+            reason: err.message,
+        },
+    };
+}
 
 /**
  * Catches errors thrown by our code and serialies them
@@ -25,33 +71,9 @@ export function errorHandler(
         return next(err);
     }
     if ((err as any).isRelayerError) {
-        const relayerError = err as RelayerBaseError;
-        if (relayerError.statusCode === HttpStatus.BAD_REQUEST) {
-            const badRequestError = relayerError as BadRequestError;
-            if (badRequestError.generalErrorCode === GeneralErrorCodes.ValidationError) {
-                const validationError = badRequestError as ValidationError;
-                const errorBody = {
-                    code: badRequestError.generalErrorCode,
-                    reason: generalErrorCodeToReason[badRequestError.generalErrorCode],
-                    validationErrors: validationError.validationErrors,
-                };
-                res.status(relayerError.statusCode).send(errorBody);
-                return;
-            } else if (badRequestError.generalErrorCode === GeneralErrorCodes.MalformedJson) {
-                const errorBody = {
-                    code: badRequestError.generalErrorCode,
-                    reason: generalErrorCodeToReason[badRequestError.generalErrorCode],
-                };
-                res.status(relayerError.statusCode).send(errorBody);
-                return;
-            }
-        } else {
-            const errorBody = {
-                reason: HttpStatus.getStatusText(relayerError.statusCode),
-            };
-            res.status(relayerError.statusCode).send(errorBody);
-            return;
-        }
+        const { statusCode, errorBody } = generateError(err);
+        res.status(statusCode).send(errorBody);
+        return;
     } else {
         return next(err);
     }
