+++ ts/src/websocket_sra.ts
@@ -1,0 +1,255 @@
+import { schemas } from '@0x/json-schemas';
+import { assetDataUtils } from '@0x/order-utils';
+import {
+    APIOrder,
+    AssetProxyId,
+    OrdersChannelMessageTypes,
+    SignedOrder,
+    UpdateOrdersChannelMessage,
+    WebsocketConnectionEventType,
+} from '@0x/types';
+import * as http from 'http';
+import * as WebSocket from 'ws';
+
+import { MalformedJSONError, NotImplementedError } from './errors';
+import { generateError } from './middleware/error_handling';
+import { utils } from './utils';
+
+export interface WebsocketSRAOpts {
+    pongInterval?: number;
+}
+
+export interface OrdersChannelSubscriptionOpts {
+    baseAssetData?: string;
+    quoteAssetData?: string;
+    makerAssetProxyId?: string;
+    takerAssetProxyId?: string;
+    makerAssetAddress?: string;
+    takerAssetAddress?: string;
+    makerAssetData?: string;
+    takerAssetData?: string;
+    traderAssetData?: string;
+    networkId?: number;
+}
+
+interface OrderChannelRequest {
+    type: string;
+    channel: MessageChannels;
+    requestId: string;
+    payload?: OrdersChannelSubscriptionOpts;
+}
+
+enum MessageTypes {
+    Subscribe = 'subscribe',
+}
+
+enum MessageChannels {
+    Orders = 'orders',
+}
+interface UpdateOrdersChannelMessageWithChannel extends UpdateOrdersChannelMessage {
+    channel: MessageChannels;
+}
+
+interface WrappedWebSocket extends WebSocket {
+    isAlive: boolean;
+    requestIds: string[];
+}
+
+const DEFAULT_OPTS: WebsocketSRAOpts = {
+    pongInterval: 5000,
+};
+
+type ALL_SUBSCRIPTION_OPTS = 'ALL_SUBSCRIPTION_OPTS';
+
+export class WebsocketSRA {
+    private readonly _server: WebSocket.Server;
+    private readonly _pongIntervalId: number;
+    private readonly _requestIdToSocket: Map<string, WrappedWebSocket> = new Map(); // requestId to WebSocket mapping
+    private readonly _requestIdToSubscriptionOpts: Map<
+        string,
+        OrdersChannelSubscriptionOpts | ALL_SUBSCRIPTION_OPTS
+    > = new Map(); // requestId -> { base, quote }
+    private static _decodedContractAndAssetData(assetData: string): { assetProxyId: string; data: string[] } {
+        let data: string[] = [assetData];
+        const decodedAssetData = assetDataUtils.decodeAssetDataOrThrow(assetData);
+        if (assetDataUtils.isMultiAssetData(decodedAssetData)) {
+            for (const nested of decodedAssetData.nestedAssetData) {
+                data = [...data, ...WebsocketSRA._decodedContractAndAssetData(nested).data];
+            }
+        } else if (assetDataUtils.isStaticCallAssetData(decodedAssetData)) {
+            // do nothing
+        } else {
+            data = [...data, decodedAssetData.tokenAddress];
+        }
+        return { data, assetProxyId: decodedAssetData.assetProxyId };
+    }
+    private static _matchesOrdersChannelSubscription(
+        order: SignedOrder,
+        opts: OrdersChannelSubscriptionOpts | ALL_SUBSCRIPTION_OPTS,
+    ): boolean {
+        if (opts === 'ALL_SUBSCRIPTION_OPTS') {
+            return true;
+        }
+        const { makerAssetData, takerAssetData } = order;
+        // Handle the specific, unambiguous asset datas
+        // traderAssetData?: string;
+        if (opts.traderAssetData && [makerAssetData, takerAssetData].includes(opts.traderAssetData)) {
+            return true;
+        }
+        // baseAssetData?: string;
+        // quoteAssetData?: string;
+        if (
+            opts.baseAssetData &&
+            opts.quoteAssetData &&
+            [makerAssetData, takerAssetData].includes(opts.baseAssetData) &&
+            [makerAssetData, takerAssetData].includes(opts.quoteAssetData)
+        ) {
+            return true;
+        }
+        // makerAssetData?: string;
+        // takerAssetData?: string;
+        if (
+            opts.makerAssetData &&
+            opts.takerAssetData &&
+            [makerAssetData, takerAssetData].includes(opts.makerAssetData) &&
+            [makerAssetData, takerAssetData].includes(opts.takerAssetData)
+        ) {
+            return true;
+        }
+        // makerAssetAddress?: string;
+        // takerAssetAddress?: string;
+        const makerContractAndAssetData = WebsocketSRA._decodedContractAndAssetData(makerAssetData);
+        const takerContractAndAssetData = WebsocketSRA._decodedContractAndAssetData(takerAssetData);
+        if (
+            opts.makerAssetAddress &&
+            opts.takerAssetAddress &&
+            makerContractAndAssetData.assetProxyId !== AssetProxyId.MultiAsset &&
+            makerContractAndAssetData.assetProxyId !== AssetProxyId.StaticCall &&
+            takerContractAndAssetData.assetProxyId !== AssetProxyId.MultiAsset &&
+            takerContractAndAssetData.assetProxyId !== AssetProxyId.StaticCall &&
+            makerContractAndAssetData.data.includes(opts.makerAssetAddress) &&
+            takerContractAndAssetData.data.includes(opts.takerAssetAddress)
+        ) {
+            return true;
+        }
+
+        // TODO (dekz)handle MAP
+        // makerAssetProxyId?: string;
+        // takerAssetProxyId?: string;
+        return false;
+    }
+    constructor(server: http.Server, opts?: WebsocketSRAOpts) {
+        const wsOpts = {
+            ...DEFAULT_OPTS,
+            ...opts,
+        };
+        this._server = new WebSocket.Server({ server });
+        this._server.on('connection', this._processConnection.bind(this));
+        this._pongIntervalId = setInterval(this._cleanupConnections.bind(this), wsOpts.pongInterval);
+    }
+    public destroy(): void {
+        clearInterval(this._pongIntervalId);
+        for (const ws of this._server.clients) {
+            ws.terminate();
+        }
+        this._requestIdToSocket.clear();
+        this._requestIdToSubscriptionOpts.clear();
+        this._server.close();
+    }
+    public orderUpdate(apiOrders: APIOrder[]): void {
+        const response: Partial<UpdateOrdersChannelMessageWithChannel> = {
+            type: OrdersChannelMessageTypes.Update,
+            channel: MessageChannels.Orders,
+            payload: apiOrders,
+        };
+        for (const order of apiOrders) {
+            const requestIdToOrders: { [requestId: string]: APIOrder[] } = {};
+            for (const [requestId, subscriptionOpts] of this._requestIdToSubscriptionOpts) {
+                if (WebsocketSRA._matchesOrdersChannelSubscription(order.order, subscriptionOpts)) {
+                    requestIdToOrders[requestId] = requestIdToOrders[requestId]
+                        ? [...requestIdToOrders[requestId], order]
+                        : [order];
+                }
+            }
+            for (const [requestId, orders] of Object.entries(requestIdToOrders)) {
+                const ws = this._requestIdToSocket.get(requestId);
+                if (ws) {
+                    ws.send(JSON.stringify({ ...response, payload: orders, requestId }));
+                }
+            }
+        }
+    }
+    private _processConnection(ws: WrappedWebSocket, _req: http.IncomingMessage): void {
+        ws.on('pong', this._pongHandler(ws).bind(this));
+        ws.on(WebsocketConnectionEventType.Message, this._messageHandler(ws).bind(this));
+        ws.on(WebsocketConnectionEventType.Close, this._closeHandler(ws).bind(this));
+        ws.isAlive = true;
+        ws.requestIds = [];
+    }
+    private _processMessage(ws: WrappedWebSocket, data: WebSocket.Data): void {
+        let message: OrderChannelRequest;
+        try {
+            message = JSON.parse(data.toString());
+        } catch (e) {
+            throw new MalformedJSONError();
+        }
+
+        utils.validateSchema(message, schemas.relayerApiOrdersChannelSubscribeSchema);
+        const requestId = message.requestId;
+        switch (message.type) {
+            case MessageTypes.Subscribe:
+                if (!message.payload) {
+                    this._requestIdToSubscriptionOpts.set(requestId, 'ALL_SUBSCRIPTION_OPTS');
+                    this._requestIdToSocket.set(requestId, ws);
+                } else {
+                    this._requestIdToSubscriptionOpts.set(requestId, message.payload);
+                    this._requestIdToSocket.set(requestId, ws);
+                }
+                break;
+            default:
+                throw new NotImplementedError(message.type);
+        }
+    }
+    private _cleanupConnections(): void {
+        // Ping every connection and if it is unresponsive
+        // terminate it during the next check
+        for (const ws of this._server.clients) {
+            if (!(ws as WrappedWebSocket).isAlive) {
+                ws.terminate();
+            } else {
+                (ws as WrappedWebSocket).isAlive = false;
+                ws.ping();
+            }
+        }
+    }
+    // tslint:disable-next-line:prefer-function-over-method
+    private _messageHandler(ws: WrappedWebSocket): (data: WebSocket.Data) => void {
+        return (data: WebSocket.Data) => {
+            try {
+                this._processMessage(ws, data);
+            } catch (err) {
+                this._processError(ws, err);
+            }
+        };
+    }
+    // tslint:disable-next-line:prefer-function-over-method
+    private _processError(ws: WrappedWebSocket, err: Error): void {
+        const { errorBody } = generateError(err);
+        ws.send(JSON.stringify(errorBody));
+        ws.terminate();
+    }
+    // tslint:disable-next-line:prefer-function-over-method
+    private _pongHandler(ws: WrappedWebSocket): () => void {
+        return () => {
+            ws.isAlive = true;
+        };
+    }
+    private _closeHandler(ws: WrappedWebSocket): () => void {
+        return () => {
+            for (const [requestId] of ws.requestIds) {
+                this._requestIdToSocket.delete(requestId);
+                this._requestIdToSubscriptionOpts.delete(requestId);
+            }
+        };
+    }
+}
